+++ src/lib.rs
@@ -1,5 +1,6 @@
-#![feature(test)]
+#![cfg_attr(test, feature(test))]
 
+#[cfg(test)]
 extern crate test;
 
 #[cfg(test)]
