+++ src/syntax/mod.rs
@@ -2,5 +2,6 @@ pub mod ast;
 pub mod parser;
 pub mod errors;
 pub mod stream;
+pub mod iter;
 
 pub use self::parser::parse;
