+++ src/syntax/stream.rs
@@ -1,126 +1,33 @@
 use super::errors::ParserError;
-
-use std::iter::Fuse;
+use super::iter::ParserStream;
 
 use std::result;
 
 type Result<T> = result::Result<T, ParserError>;
 
-#[derive(Clone, Debug)]
-pub struct ParserStream<I>
-    where I: Iterator
-{
-    iter: Fuse<I>,
-    pub buf: Vec<char>,
-    peek_index: i32,
-    index: i32,
+pub trait FTLParserStream<I> {
+    fn peek_line_ws(&mut self);
+    fn skip_ws_lines(&mut self);
+    fn skip_line_ws(&mut self);
+    fn skip_ws(&mut self);
+    fn expect_char(&mut self, ch: char) -> Result<()>;
+    fn take_char_if(&mut self, ch: char) -> bool;
+    fn take_char_after_line_ws_if(&mut self, ch2: char) -> bool;
 
-    ch: Option<char>,
+    fn take_char<F>(&mut self, f: F) -> Option<char> where F: Fn(char) -> bool;
+    fn current_char_matches<F>(&mut self, f: F) -> bool where F: Fn(char) -> bool;
+    fn peek_char_matches<F>(&mut self, f: F) -> bool where F: Fn(char) -> bool;
 
-    iter_end: bool,
-    peek_end: bool,
+    fn is_id_start(&mut self) -> bool;
+    fn take_id_start(&mut self) -> Result<char>;
+    fn take_id_char(&mut self) -> Option<char>;
+    fn take_kw_char(&mut self) -> Option<char>;
 }
 
-impl<I: Iterator<Item = char>> ParserStream<I> {
-    pub fn current(&mut self) -> Option<char> {
-        self.ch
-    }
-
-    pub fn current_is(&mut self, ch: char) -> bool {
-        self.ch == Some(ch)
-    }
-
-    pub fn current_peek(&self) -> Option<char> {
-        let diff = self.peek_index - self.index;
-
-        if diff == 0 {
-            return self.ch;
-        }
-
-        if self.peek_end {
-            return None;
-        }
-
-        return Some(self.buf[(diff - 1) as usize]);
-    }
-
-    pub fn current_peek_is(&mut self, ch: char) -> bool {
-        match self.current_peek() {
-            Some(c) => ch == c,
-            None => false,
-        }
-    }
-
-    // pub fn bump(&mut self) {
-    //     self.peek_index = None;
-    //     if self.buf.is_empty() {
-    //         self.iter.next();
-    //     } else {
-    //         self.buf.remove(0);
-    //     }
-    // }
-
-    pub fn peek(&mut self) -> Option<char> {
-        if !self.peek_end {
-            self.peek_index += 1;
-        }
-        match self.iter.next() {
-            Some(c) => {
-                self.buf.push(c);
-                let diff = (self.peek_index - self.index) as usize;
-                return Some(self.buf[diff - 1]);
-            }
-            None => {
-                self.peek_end = true;
-                return None;
-            }
-        }
-    }
-
-    pub fn get_index(&self) -> i32 {
-        self.index
-    }
-
-    pub fn get_peek_index(&self) -> i32 {
-        return self.peek_index;
-    }
-
-    pub fn has_more(&mut self) -> bool {
-        let ret = self.peek().is_some();
-
-        self.reset_peek();
-
-        ret
-    }
-
-    pub fn peek_char_is(&mut self, c: char) -> bool {
-        if self.peek_end {
-            return false;
-        }
-        let ret = match self.peek() {
-            Some(ch) if ch == c => true,
-            _ => false,
-        };
-
-        self.peek_index -= 1;
-        return ret;
-    }
-
-    pub fn reset_peek(&mut self) {
-        self.peek_index = self.index;
-    }
-
-    pub fn skip_to_peek(&mut self) {
-        let diff = self.peek_index - self.index;
-
-        for _ in 0..diff {
-            self.ch = Some(self.buf.remove(0));
-        }
-
-        self.index = self.peek_index;
-    }
-
-    pub fn peek_line_ws(&mut self) {
+impl<I> FTLParserStream<I> for ParserStream<I>
+    where I: Iterator<Item = char>
+{
+    fn peek_line_ws(&mut self) {
         while let Some(ch) = self.peek() {
             if ch != ' ' && ch != '\t' {
                 break;
@@ -128,7 +35,7 @@ impl<I: Iterator<Item = char>> ParserStream<I> {
         }
     }
 
-    pub fn skip_ws_lines(&mut self) {
+    fn skip_ws_lines(&mut self) {
         loop {
             self.peek_line_ws();
 
@@ -145,7 +52,7 @@ impl<I: Iterator<Item = char>> ParserStream<I> {
         }
     }
 
-    pub fn skip_line_ws(&mut self) {
+    fn skip_line_ws(&mut self) {
         while let Some(ch) = self.ch {
             if ch != ' ' && ch != '\t' {
                 break;
@@ -155,7 +62,7 @@ impl<I: Iterator<Item = char>> ParserStream<I> {
         }
     }
 
-    pub fn skip_ws(&mut self) {
+    fn skip_ws(&mut self) {
         while let Some(ch) = self.ch {
             if ch != ' ' && ch != '\n' && ch != '\t' && ch != '\r' {
                 break;
@@ -165,7 +72,7 @@ impl<I: Iterator<Item = char>> ParserStream<I> {
         }
     }
 
-    pub fn expect_char(&mut self, ch: char) -> Result<()> {
+    fn expect_char(&mut self, ch: char) -> Result<()> {
         match self.ch {
             Some(ch2) if ch == ch2 => {
                 self.next();
@@ -175,7 +82,7 @@ impl<I: Iterator<Item = char>> ParserStream<I> {
         }
     }
 
-    pub fn take_char_if(&mut self, ch: char) -> bool {
+    fn take_char_if(&mut self, ch: char) -> bool {
         match self.ch {
             Some(ch2) if ch == ch2 => {
                 self.next();
@@ -185,7 +92,7 @@ impl<I: Iterator<Item = char>> ParserStream<I> {
         }
     }
 
-    pub fn take_char_after_line_ws_if(&mut self, ch2: char) -> bool {
+    fn take_char_after_line_ws_if(&mut self, ch2: char) -> bool {
         while let Some(ch) = self.current_peek() {
             if ch == ' ' || ch == '\t' {
                 self.peek();
@@ -203,7 +110,7 @@ impl<I: Iterator<Item = char>> ParserStream<I> {
         return false;
     }
 
-    pub fn take_char<F>(&mut self, f: F) -> Option<char>
+    fn take_char<F>(&mut self, f: F) -> Option<char>
         where F: Fn(char) -> bool
     {
 
@@ -216,7 +123,7 @@ impl<I: Iterator<Item = char>> ParserStream<I> {
         }
     }
 
-    pub fn current_char_matches<F>(&mut self, f: F) -> bool
+    fn current_char_matches<F>(&mut self, f: F) -> bool
         where F: Fn(char) -> bool
     {
 
@@ -226,7 +133,7 @@ impl<I: Iterator<Item = char>> ParserStream<I> {
         }
     }
 
-    pub fn peek_char_matches<F>(&mut self, f: F) -> bool
+    fn peek_char_matches<F>(&mut self, f: F) -> bool
         where F: Fn(char) -> bool
     {
 
@@ -242,7 +149,7 @@ impl<I: Iterator<Item = char>> ParserStream<I> {
         }
     }
 
-    pub fn is_id_start(&mut self) -> bool {
+    fn is_id_start(&mut self) -> bool {
         let closure = |x| match x {
             'a'...'z' | 'A'...'Z' | '_' => true,
             _ => false,
@@ -251,7 +158,7 @@ impl<I: Iterator<Item = char>> ParserStream<I> {
         return self.current_char_matches(closure);
     }
 
-    pub fn take_id_start(&mut self) -> Result<char> {
+    fn take_id_start(&mut self) -> Result<char> {
         let closure = |x| match x {
             'a'...'z' | 'A'...'Z' | '_' => true,
             _ => false,
@@ -267,7 +174,7 @@ impl<I: Iterator<Item = char>> ParserStream<I> {
         }
     }
 
-    pub fn take_id_char(&mut self) -> Option<char> {
+    fn take_id_char(&mut self) -> Option<char> {
         let closure = |x| match x {
             'a'...'z' | 'A'...'Z' | '0'...'9' | '_' | '-' => true,
             _ => false,
@@ -279,7 +186,7 @@ impl<I: Iterator<Item = char>> ParserStream<I> {
         }
     }
 
-    pub fn take_kw_char(&mut self) -> Option<char> {
+    fn take_kw_char(&mut self) -> Option<char> {
         let closure = |x| match x {
             'a'...'z' | 'A'...'Z' | '0'...'9' | '_' | '-' | ' ' => true,
             _ => false,
@@ -289,48 +196,5 @@ impl<I: Iterator<Item = char>> ParserStream<I> {
             Some(ch) => Some(ch),
             None => None,
         }
-    }
-}
-
-impl<I> Iterator for ParserStream<I>
-    where I: Iterator<Item = char>
-{
-    type Item = char;
-
-    fn next(&mut self) -> Option<char> {
-
-        if self.buf.is_empty() {
-            self.ch = self.iter.next();
-        } else {
-            self.ch = Some(self.buf.remove(0));
-        }
-
-        if !self.iter_end {
-            self.index += 1;
-        }
-
-        if self.ch.is_none() {
-            self.iter_end = true;
-            self.peek_end = true;
-        }
-
-        self.peek_index = self.index;
-
-        self.ch
-    }
-}
-
-pub fn parserstream<I>(iterable: I) -> ParserStream<I::IntoIter>
-    where I: IntoIterator
-{
-    ParserStream {
-        iter: iterable.into_iter().fuse(),
-        buf: vec![],
-        peek_index: -1,
-        index: -1,
-        ch: None,
-
-        iter_end: false,
-        peek_end: false,
     }
 }
